order/login/register

+++ src/main/java/com/springboot/jewellerysystem/controller/BannerImageController.java
@@ -20,6 +20,7 @@ import com.springboot.jewellerysystem.entity.BannerImage;
 import com.springboot.jewellerysystem.service.BannerImageService;
 import com.springboot.jewellerysystem.service.BannerService;
 import com.springboot.jewellerysystem.util.FileUploadUtil;
+import com.springboot.jewellerysystem.util.Helper;
 
 
 
@@ -36,6 +37,9 @@ public class BannerImageController {
 
 	@GetMapping(value = "/index")
 	public String bannerImages(Model model, @RequestParam(name = "keyword", defaultValue = "") String keyword) {
+		if(Helper.checkUserRole()) { return "redirect:/";}
+		if(!Helper.checkAdminRole()) {return "redirect:/admin/logout";}
+				
 		List<BannerImage> bannerImages = bannerImageService.getAllBannerImage();
 		model.addAttribute("listBannerImages", bannerImages);
 		model.addAttribute("keyword", keyword);
@@ -44,6 +48,9 @@ public class BannerImageController {
 
 	@GetMapping(value = "/create")
 	public String formBannerImages(Model model) {
+		if(Helper.checkUserRole()) { return "redirect:/";}
+		if(!Helper.checkAdminRole()) {return "redirect:/admin/logout";}
+				
 		model.addAttribute("bannerImage", new BannerImage());
 		List<Banner> banners = bannerService.getAllBanner();
 		model.addAttribute("listBanners", banners);
@@ -53,6 +60,9 @@ public class BannerImageController {
 
 	@GetMapping(value = "/delete/{id}")
 	public String deleteBannerImage(@PathVariable(value = "id") Integer id, String keyword,HttpSession session) {
+		if(Helper.checkUserRole()) { return "redirect:/";}
+		if(!Helper.checkAdminRole()) {return "redirect:/admin/logout";}
+				
 		bannerImageService.removeBannerImage(id);
 		session.setAttribute("msg", "deleted");
 		return "redirect:/admin/bannerImage/index?keyword=" + keyword;
@@ -60,6 +70,9 @@ public class BannerImageController {
 
 	@GetMapping(value = "/update/{id}")
 	public String updateBannerImage(@PathVariable(value = "id") Integer id, Model model) {
+		if(Helper.checkUserRole()) { return "redirect:/";}
+		if(!Helper.checkAdminRole()) {return "redirect:/admin/logout";}
+				
 		BannerImage bannerImage = bannerImageService.loadBannerImageById(id);
 		
 		model.addAttribute("bannerImage", bannerImage);
