+++ src/main/java/com/springboot/jewellerysystem/controller/CartController.java
@@ -18,6 +18,7 @@ import com.springboot.jewellerysystem.entity.User;
 import com.springboot.jewellerysystem.service.CartService;
 import com.springboot.jewellerysystem.service.ProductService;
 import com.springboot.jewellerysystem.service.UserService;
+import com.springboot.jewellerysystem.util.Helper;
 
 @Controller
 @RequestMapping(value = "admin/cart")
@@ -34,6 +35,8 @@ public class CartController {
 
 	@GetMapping(value = "/index")
 	public String carts(Model model, @RequestParam(name = "keyword", defaultValue = "") String keyword) {
+		if(Helper.checkUserRole()) { return "redirect:/";}
+    	if(!Helper.checkAdminRole()) {return "redirect:/admin/logout";}
 		List<Cart> carts = cartService.getAllCart();
 		model.addAttribute("listCarts", carts);
 		model.addAttribute("keyword", keyword);
@@ -42,6 +45,8 @@ public class CartController {
 
 	@GetMapping(value = "/create")
 	public String formCarts(Model model) {
+		if(Helper.checkUserRole()) { return "redirect:/";}
+    	if(!Helper.checkAdminRole()) {return "redirect:/admin/logout";}
 		model.addAttribute("cart", new Cart());
 		List<User> users = userService.getAllUser();
 		model.addAttribute("listUsers", users);
@@ -54,6 +59,8 @@ public class CartController {
 
 	@GetMapping(value = "/delete/{id}")
 	public String deleteCart(@PathVariable(value = "id") Integer id, String keyword, HttpSession session) {
+		if(Helper.checkUserRole()) { return "redirect:/";}
+    	if(!Helper.checkAdminRole()) {return "redirect:/admin/logout";}
 		cartService.removeCart(id);
 		session.setAttribute("msg", "deleted");
 		return "redirect:/admin/cart/index?keyword=" + keyword;
@@ -61,6 +68,8 @@ public class CartController {
 
 	@GetMapping(value = "/update/{id}")
 	public String updateCart(@PathVariable(value = "id") Integer id, Model model) {
+		if(Helper.checkUserRole()) { return "redirect:/";}
+    	if(!Helper.checkAdminRole()) {return "redirect:/admin/logout";}
 		Cart cart = cartService.loadCartById(id);
 		model.addAttribute("cart", cart);
 		List<User> users = userService.getAllUser();
