+++ src/main/java/com/springboot/jewellerysystem/controller/ContactUsController.java
@@ -14,7 +14,8 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 
 import com.springboot.jewellerysystem.entity.ContactUs;
-import com.springboot.jewellerysystem.service.ContactUsService; 
+import com.springboot.jewellerysystem.service.ContactUsService;
+import com.springboot.jewellerysystem.util.Helper; 
 @Controller 
 @RequestMapping(value = "admin/contactUs") 
 public class ContactUsController { 
@@ -25,7 +26,9 @@ public class ContactUsController {
  
     @GetMapping(value = "/index") 
     public String contactUses(Model model, @RequestParam(name = "keyword", defaultValue = "") String keyword) { 
-        List<ContactUs> contactUses = contactUsService.getAllContactUs(); 
+    	if(Helper.checkUserRole()) { return "redirect:/";}
+    	if(!Helper.checkAdminRole()) {return "redirect:/admin/logout";}
+    	List<ContactUs> contactUses = contactUsService.getAllContactUs(); 
         model.addAttribute("listContactUses", contactUses); 
         model.addAttribute("keyword", keyword); 
         return "admin/list/contactUses_list"; 
@@ -33,20 +36,26 @@ public class ContactUsController {
  
   @GetMapping(value = "/create") 
     public String formContactUses(Model model) { 
-        model.addAttribute("contactUs", new ContactUs()); 
+	  if(Helper.checkUserRole()) { return "redirect:/";}
+  	if(!Helper.checkAdminRole()) {return "redirect:/admin/logout";}  
+	  model.addAttribute("contactUs", new ContactUs()); 
         model.addAttribute("myDate", new Date());
         return "admin/entry/contactUs_entry"; 
     } 
     @GetMapping(value = "/delete/{id}") 
     public String deleteContactUs(@PathVariable(value = "id") Integer id, String keyword, HttpSession session) { 
-        contactUsService.removeContactUs(id); 
+    	if(Helper.checkUserRole()) { return "redirect:/";}
+    	if(!Helper.checkAdminRole()) {return "redirect:/admin/logout";}
+    	contactUsService.removeContactUs(id); 
         session.setAttribute("msg", "deleted");
         return "redirect:/admin/contactUs/index?keyword=" + keyword; 
     }
  
     @GetMapping(value = "/update/{id}") 
     public String updateContactUs(@PathVariable(value = "id") Integer id, Model model) { 
-        ContactUs contactUs = contactUsService.loadContactUsById(id); 
+    	if(Helper.checkUserRole()) { return "redirect:/";}
+    	if(!Helper.checkAdminRole()) {return "redirect:/admin/logout";}
+    	ContactUs contactUs = contactUsService.loadContactUsById(id); 
         model.addAttribute("contactUs", contactUs); 
         return "admin/edit/contactUs_edit"; 
     }
