+++ src/main/java/com/springboot/jewellerysystem/controller/CountryController.java
@@ -1,7 +1,9 @@
 package com.springboot.jewellerysystem.controller;
 
 import com.springboot.jewellerysystem.entity.Country; 
-import com.springboot.jewellerysystem.service.CountryService; 
+import com.springboot.jewellerysystem.service.CountryService;
+import com.springboot.jewellerysystem.util.Helper;
+
 import org.springframework.stereotype.Controller; 
 import org.springframework.ui.Model; 
 import org.springframework.web.bind.annotation.GetMapping; 
@@ -22,7 +24,9 @@ public class CountryController {
  
     @GetMapping(value = "/index") 
     public String countries(Model model, @RequestParam(name = "keyword", defaultValue = "") String keyword) { 
-        List<Country> countries = countryService.getAllCountry(); 
+    	if(Helper.checkUserRole()) { return "redirect:/";}
+    	if(!Helper.checkAdminRole()) {return "redirect:/admin/logout";}
+    	List<Country> countries = countryService.getAllCountry(); 
         model.addAttribute("listCountries", countries); 
         model.addAttribute("keyword", keyword); 
         return "admin/list/countries_list"; 
@@ -30,19 +34,25 @@ public class CountryController {
  
   @GetMapping(value = "/create") 
     public String formCountries(Model model) { 
-        model.addAttribute("country", new Country()); 
+	  if(Helper.checkUserRole()) { return "redirect:/";}
+  	if(!Helper.checkAdminRole()) {return "redirect:/admin/logout";} 
+	  model.addAttribute("country", new Country()); 
         return "admin/entry/country_entry"; 
     } 
     @GetMapping(value = "/delete/{id}") 
     public String deleteCountry(@PathVariable(value = "id") Integer id, String keyword, HttpSession session) { 
-        countryService.removeCountry(id); 
+    	if(Helper.checkUserRole()) { return "redirect:/";}
+    	if(!Helper.checkAdminRole()) {return "redirect:/admin/logout";}
+    	countryService.removeCountry(id); 
         session.setAttribute("msg", "deleted");
         return "redirect:/admin/country/index?keyword=" + keyword; 
     }
  
     @GetMapping(value = "/update/{id}") 
     public String updateCountry(@PathVariable(value = "id") Integer id, Model model) { 
-        Country country = countryService.loadCountryById(id); 
+    	if(Helper.checkUserRole()) { return "redirect:/";}
+    	if(!Helper.checkAdminRole()) {return "redirect:/admin/logout";}
+    	Country country = countryService.loadCountryById(id); 
         model.addAttribute("country", country); 
         return "admin/edit/country_edit"; 
     }
