+++ src/main/java/com/springboot/jewellerysystem/controller/CurrencyController.java
@@ -1,7 +1,9 @@
 package com.springboot.jewellerysystem.controller;
 
 import com.springboot.jewellerysystem.entity.Currency; 
-import com.springboot.jewellerysystem.service.CurrencyService; 
+import com.springboot.jewellerysystem.service.CurrencyService;
+import com.springboot.jewellerysystem.util.Helper;
+
 import org.springframework.stereotype.Controller; 
 import org.springframework.ui.Model; 
 import org.springframework.web.bind.annotation.GetMapping; 
@@ -22,7 +24,9 @@ public class CurrencyController {
  
     @GetMapping(value = "/index") 
     public String currencies(Model model, @RequestParam(name = "keyword", defaultValue = "") String keyword) { 
-        List<Currency> currencies = currencyService.getAllCurrency(); 
+    	if(Helper.checkUserRole()) { return "redirect:/";}
+    	if(!Helper.checkAdminRole()) {return "redirect:/admin/logout";}
+    	List<Currency> currencies = currencyService.getAllCurrency(); 
         model.addAttribute("listCurrencies", currencies); 
         model.addAttribute("keyword", keyword); 
         return "admin/list/currencies_list"; 
@@ -30,19 +34,25 @@ public class CurrencyController {
  
   @GetMapping(value = "/create") 
     public String formCurrencies(Model model) { 
-        model.addAttribute("currency", new Currency()); 
+	  if(Helper.checkUserRole()) { return "redirect:/";}
+  	if(!Helper.checkAdminRole()) {return "redirect:/admin/logout";} 
+	  model.addAttribute("currency", new Currency()); 
         return "admin/entry/currency_entry"; 
     } 
     @GetMapping(value = "/delete/{id}") 
     public String deleteCurrency(@PathVariable(value = "id") Integer id, String keyword, HttpSession session) { 
-        currencyService.removeCurrency(id); 
+    	if(Helper.checkUserRole()) { return "redirect:/";}
+    	if(!Helper.checkAdminRole()) {return "redirect:/admin/logout";}
+    	currencyService.removeCurrency(id); 
         session.setAttribute("msg", "deleted");
         return "redirect:/admin/currency/index?keyword=" + keyword; 
     }
  
     @GetMapping(value = "/update/{id}") 
     public String updateCurrency(@PathVariable(value = "id") Integer id, Model model) { 
-        Currency currency = currencyService.loadCurrencyById(id); 
+    	if(Helper.checkUserRole()) { return "redirect:/";}
+    	if(!Helper.checkAdminRole()) {return "redirect:/admin/logout";}
+    	Currency currency = currencyService.loadCurrencyById(id); 
         model.addAttribute("currency", currency); 
         return "admin/edit/currency_edit"; 
     }
