+++ src/main/java/com/springboot/jewellerysystem/controller/FeedbackController.java
@@ -1,7 +1,9 @@
 package com.springboot.jewellerysystem.controller;
 
 import com.springboot.jewellerysystem.entity.Feedback; 
-import com.springboot.jewellerysystem.service.FeedbackService; 
+import com.springboot.jewellerysystem.service.FeedbackService;
+import com.springboot.jewellerysystem.util.Helper;
+
 import org.springframework.stereotype.Controller; 
 import org.springframework.ui.Model; 
 import org.springframework.web.bind.annotation.GetMapping; 
@@ -24,7 +26,9 @@ public class FeedbackController {
  
     @GetMapping(value = "/index") 
     public String feedbacks(Model model, @RequestParam(name = "keyword", defaultValue = "") String keyword) { 
-        List<Feedback> feedbacks = feedbackService.getAllFeedback(); 
+    	if(Helper.checkUserRole()) { return "redirect:/";}
+    	if(!Helper.checkAdminRole()) {return "redirect:/admin/logout";}
+    	List<Feedback> feedbacks = feedbackService.getAllFeedback(); 
         model.addAttribute("listFeedbacks", feedbacks); 
         model.addAttribute("keyword", keyword); 
         return "admin/list/feedbacks_list"; 
@@ -32,18 +36,24 @@ public class FeedbackController {
  
   @GetMapping(value = "/create") 
     public String formFeedbacks(Model model) { 
-        model.addAttribute("feedback", new Feedback()); 
+	  if(Helper.checkUserRole()) { return "redirect:/";}
+  	if(!Helper.checkAdminRole()) {return "redirect:/admin/logout";}  
+	  model.addAttribute("feedback", new Feedback()); 
         return "admin/entry/feedback_entry"; 
     } 
     @GetMapping(value = "/delete/{id}") 
     public String deleteFeedback(@PathVariable(value = "id") Integer id, String keyword, HttpSession session) { 
-        feedbackService.removeFeedback(id); 
+    	if(Helper.checkUserRole()) { return "redirect:/";}
+    	if(!Helper.checkAdminRole()) {return "redirect:/admin/logout";}
+    	feedbackService.removeFeedback(id); 
         return "redirect:/admin/feedback/index?keyword=" + keyword; 
     }
  
     @GetMapping(value = "/update/{id}") 
     public String updateFeedback(@PathVariable(value = "id") Integer id, Model model) { 
-        Feedback feedback = feedbackService.loadFeedbackById(id); 
+    	if(Helper.checkUserRole()) { return "redirect:/";}
+    	if(!Helper.checkAdminRole()) {return "redirect:/admin/logout";}
+    	Feedback feedback = feedbackService.loadFeedbackById(id); 
         model.addAttribute("feedback", feedback); 
         return "admin/edit/feedback_edit"; 
     }
