+++ src/main/java/com/springboot/jewellerysystem/controller/LinkController.java
@@ -1,7 +1,9 @@
 package com.springboot.jewellerysystem.controller;
 
 import com.springboot.jewellerysystem.entity.Link; 
-import com.springboot.jewellerysystem.service.LinkService; 
+import com.springboot.jewellerysystem.service.LinkService;
+import com.springboot.jewellerysystem.util.Helper;
+
 import org.springframework.stereotype.Controller; 
 import org.springframework.ui.Model; 
 import org.springframework.web.bind.annotation.GetMapping; 
@@ -22,7 +24,9 @@ public class LinkController {
  
     @GetMapping(value = "/index") 
     public String links(Model model, @RequestParam(name = "keyword", defaultValue = "") String keyword) { 
-        List<Link> links = linkService.getAllLink(); 
+    	if(Helper.checkUserRole()) { return "redirect:/";}
+    	if(!Helper.checkAdminRole()) {return "redirect:/admin/logout";}
+    	List<Link> links = linkService.getAllLink(); 
         model.addAttribute("listLinks", links); 
         model.addAttribute("keyword", keyword); 
         return "admin/list/links_list"; 
@@ -30,19 +34,25 @@ public class LinkController {
  
   @GetMapping(value = "/create") 
     public String formLinks(Model model) { 
-        model.addAttribute("link", new Link()); 
+	  if(Helper.checkUserRole()) { return "redirect:/";}
+  	if(!Helper.checkAdminRole()) {return "redirect:/admin/logout";} 
+	  model.addAttribute("link", new Link()); 
         return "admin/entry/link_entry"; 
     } 
     @GetMapping(value = "/delete/{id}") 
     public String deleteLink(@PathVariable(value = "id") Integer id, String keyword, HttpSession session) { 
-        linkService.removeLink(id); 
+    	if(Helper.checkUserRole()) { return "redirect:/";}
+    	if(!Helper.checkAdminRole()) {return "redirect:/admin/logout";}
+    	linkService.removeLink(id); 
         session.setAttribute("msg", "deleted");
         return "redirect:/admin/link/index?keyword=" + keyword; 
     }
  
     @GetMapping(value = "/update/{id}") 
     public String updateLink(@PathVariable(value = "id") Integer id, Model model) { 
-        Link link = linkService.loadLinkById(id); 
+    	if(Helper.checkUserRole()) { return "redirect:/";}
+    	if(!Helper.checkAdminRole()) {return "redirect:/admin/logout";}
+    	Link link = linkService.loadLinkById(id); 
         model.addAttribute("link", link); 
         return "admin/edit/link_edit"; 
     }
